Prevent cqerl from crashing when a client fails to start
On network failure, the following chain of events may happen:
- Client supervisor restart intensity is exhausted because the
network stays broken
- The whole pool is restarted
- Some clients connect, at least one of them it fails to start
  (the network is broken again)
- The pool supervisor is shut down with 'failed_to_start_child'
- cqerl_hash failed to create the client table, but
  it still starts receiving the 'add_client' messages
  from clients that connected but are now already dead
- Handling the stale 'add_client' causes cqerl_hash to crash
  with 'badmatch'
- All children of 'cqerl_sup' are restarted because of 'one_for_all'
- If this happens a few times in a row, cqerl app goes down
  and the whole Erlang node crashes.

The simplest change needed to prevent this escalation is to just
ignore the stale 'client_started' notifications from clients that are
already dead.

This is confirmed to work by 'mongoose_cassandra_SUITE' in MongooseIM.
The test was run many times with simulated network issues.
It was crashing all the time before the change and does not crash at
all after the change. The supervision tree is never restarted as well.